+++ app/Exceptions/Video/VideoException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Exceptions\Video;
+
+use RuntimeException;
+
+abstract class VideoException extends RuntimeException
+{
+}